+++ src/service/CityService.js
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+export const getAllCities = async () => {
+    return await axios.get("/cities").then((response) => {
+        return response.data
+    });
+};
